refactor(clawdstrike-policy-event): extract projection from edr/mod.rs
Move the privacy-projection cluster (project_observation_privacy,
project_event_privacy, count_projection_class, the push_optional_*/
push_metadata/push_hash/push_local family, and privacy_feature) into
edr/projection.rs.

privacy.rs reaches project_observation_privacy and count_projection_class
via super::, so those two are re-exported pub(crate) at the edr root; the
push_* helpers and privacy_feature stay private to projection. The
projection fns reach util (normalize_path_string) and the edr telemetry
types via super::. Pure relocation; bodies unchanged.
